+++ lib/peddler/errors/error.rb
@@ -1,5 +1,14 @@
 module Peddler
+  # @api private
   module Errors
+    KNOWN = %w(
+      AccessDenied
+      InvalidMarketplace
+      InvalidParameterValue
+      QuotaExceeded
+      RequestThrottled
+    ).freeze
+
     # @api private
     class Error < StandardError
       attr_reader :cause
@@ -8,6 +17,10 @@ module Peddler
         @cause = cause
         super msg
       end
+    end
+
+    KNOWN.each do |name|
+      const_set name, Class.new(Error)
     end
   end
 end