+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ConfirmPreorder, GetPrepInstructionsForSKU, GetPrepInstructionsForASIN,
 GetUniquePackageLabels, and GetPalletLabels operations to the
 FullfillmentInboundShipment client.
+- Add known error types to Peddler::Errors.
 
 ### Fixed
 - Don't camelise symbol keys with capital letters.